Better Registers
Refactored `Regester` so that there is simple enum of general purpose registers.
`Register` is now a struct that has a GPR, register width, and an optional `Byte` field.
The latter is for `AH`, `AL`, etc.

This should play better with upstream users, since the registers are now independed of their widths.
It also made everything in the lib cleaner.